+++ src/components/shell/index.ts
@@ -1,0 +1,7 @@
+export { PrimarySidebar, type SidebarSession } from './PrimarySidebar'
+export {
+  SecondaryNav,
+  shouldShowSecondaryNav,
+  type ShellView,
+} from './SecondaryNav'
+export { formatSessionStamp, formatSessionPreview } from './session-format'